requirements.txt
numpy==1.24.3
pygame==2.4.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
